moved get_msearch_query to utils
